+++ test/specs/Plugin.spec.js
@@ -1,10 +1,12 @@
 import path from 'path'
 import consola from 'consola'
-import { serial as test, before, after } from 'ava'
+import { serial as test } from 'ava'
 import config from '@/nuxt.config'
 import { state as indexState } from '@/store/index'
 import { state as examplesState } from '@/store/examples'
-import { compilePlugin, ioServerInit, removeCompiledPlugin } from '@/test/utils'
+import { compileAndImportPlugin, removeCompiledPlugin } from '@/test/utils'
+
+import Plugin, { pOptions } from '@/io/plugin.compiled'
 
 const { io } = config
 const state = indexState()
@@ -13,22 +15,15 @@ state.examples.__ob__ = ''
 const src = path.resolve('./io/plugin.js')
 const tmpFile = path.resolve('./io/plugin.compiled.js')
 
-async function compile(t) {
-  const compiled = await compilePlugin({ src, tmpFile, options: io }).catch(
-    (err) => {
-      consola.error(err)
-      t.fail()
-    }
-  )
-  Plugin = compiled.Plugin
-  pOptions = compiled.pOptions
-  t.pass()
-}
-
-function loadPlugin(t, ioOpts = {}) {
+function loadPlugin({
+  t,
+  ioOpts = {},
+  plugin = Plugin,
+  callCnt = { storeWatch: 0 }
+}) {
   return new Promise((resolve, reject) => {
     const context = {}
-    Plugin(context, (label, NuxtSocket) => {
+    plugin(context, (label, NuxtSocket) => {
       context.$store = {
         commit: (msg) => {
           consola.log('commit', msg)
@@ -37,6 +32,7 @@ function loadPlugin(t, ioOpts = {}) {
           consola.log('dispatch', msg)
         },
         watch: (stateCb, dataCb) => {
+          callCnt.storeWatch++
           stateCb(state)
           dataCb({ sample: 123 })
         }
@@ -57,20 +53,10 @@ function loadPlugin(t, ioOpts = {}) {
   })
 }
 
-let Plugin, pOptions
-
-before('Compile Plugin', compile)
-
-before('Init IO Server', ioServerInit)
-
-after('Remove compiled plugin', () => {
-  removeCompiledPlugin(tmpFile)
-})
-
 test('Socket plugin (empty options)', async (t) => {
   const testCfg = { sockets: [] }
   pOptions.set(testCfg)
-  await loadPlugin(t).catch((e) => {
+  await loadPlugin({ t }).catch((e) => {
     t.is(
       e.message,
       "Please configure sockets if planning to use nuxt-socket-io: \r\n [{name: '', url: ''}]"
@@ -81,7 +67,7 @@ test('Socket plugin (empty options)', async (t) => {
 test('Socket plugin (malformed sockets)', async (t) => {
   const testCfg = { sockets: {} }
   pOptions.set(testCfg)
-  await loadPlugin(t).catch((e) => {
+  await loadPlugin({ t }).catch((e) => {
     t.is(
       e.message,
       "Please configure sockets if planning to use nuxt-socket-io: \r\n [{name: '', url: ''}]"
@@ -92,7 +78,7 @@ test('Socket plugin (malformed sockets)', async (t) => {
 test('Socket plugin (options missing info)', async (t) => {
   const testCfg = { sockets: [{}] }
   pOptions.set(testCfg)
-  await loadPlugin(t).catch((e) => {
+  await loadPlugin({ t }).catch((e) => {
     t.is(e.message, 'URL must be defined for nuxtSocket')
   })
 })
@@ -108,7 +94,7 @@ test('Socket plugin (no vuex options)', async (t) => {
     ]
   }
   pOptions.set(testCfg)
-  await loadPlugin(t, { name: 'home' })
+  await loadPlugin({ t, ioOpts: { name: 'home' } })
 })
 
 test('Socket plugin (malformed vuex options)', async (t) => {
@@ -126,7 +112,7 @@ test('Socket plugin (malformed vuex options)', async (t) => {
     ]
   }
   pOptions.set(testCfg)
-  await loadPlugin(t)
+  await loadPlugin({ t })
 })
 
 test('Socket plugin (vuex options missing mutations)', async (t) => {
@@ -142,7 +128,7 @@ test('Socket plugin (vuex options missing mutations)', async (t) => {
     ]
   }
   pOptions.set(testCfg)
-  await loadPlugin(t)
+  await loadPlugin({ t })
 })
 
 test('Socket plugin (vuex options missing actions)', async (t) => {
@@ -158,7 +144,7 @@ test('Socket plugin (vuex options missing actions)', async (t) => {
     ]
   }
   pOptions.set(testCfg)
-  await loadPlugin(t)
+  await loadPlugin({ t })
 })
 
 test('Socket plugin (vuex opts as strings)', async (t) => {
@@ -175,7 +161,7 @@ test('Socket plugin (vuex opts as strings)', async (t) => {
     ]
   }
   pOptions.set(testCfg)
-  await loadPlugin(t)
+  await loadPlugin({ t })
 })
 
 test('Socket plugin (malformed emitBacks)', async (t) => {
@@ -194,7 +180,7 @@ test('Socket plugin (malformed emitBacks)', async (t) => {
     ]
   }
   pOptions.set(testCfg)
-  await loadPlugin(t).catch((e) => {
+  await loadPlugin({ t }).catch((e) => {
     t.is(
       e.message,
       emitBack + ' is a vuex module. You probably want to watch its properties'
@@ -202,25 +188,122 @@ test('Socket plugin (malformed emitBacks)', async (t) => {
   })
 })
 
+test('Emitback is not defined in vuex store', async (t) => {
+  const errEmitBack = 'something/undefined'
+  const testCfg = {
+    sockets: [
+      {
+        default: true,
+        url: 'http://localhost:3000',
+        vuex: {
+          actions: [],
+          mutations: [],
+          emitBacks: [errEmitBack]
+        }
+      }
+    ]
+  }
+  pOptions.set(testCfg)
+  await loadPlugin({ t }).catch((e) => {
+    t.is(
+      e.message,
+      [
+        `[nuxt-socket-io]: Trying to register emitback ${errEmitBack} failed`,
+        `because it is not defined in Vuex.`,
+        'Is state set up correctly in your stores folder?'
+      ].join('\n')
+    )
+  })
+})
+
+test('Emitback is not defined in vuex store (variant 2)', async (t) => {
+  const errEmitBack = { 'something/undefined': 'someData' }
+  const [errEmitBackLabel] = Object.keys(errEmitBack)
+  const testCfg = {
+    sockets: [
+      {
+        default: true,
+        url: 'http://localhost:3000',
+        vuex: {
+          actions: [],
+          mutations: [],
+          emitBacks: [errEmitBack]
+        }
+      }
+    ]
+  }
+  pOptions.set(testCfg)
+  await loadPlugin({ t }).catch((e) => {
+    t.is(
+      e.message,
+      [
+        `[nuxt-socket-io]: Trying to register emitback ${errEmitBackLabel} failed`,
+        `because it is not defined in Vuex.`,
+        'Is state set up correctly in your stores folder?'
+      ].join('\n')
+    )
+  })
+})
+
+test('Duplicate Watchers are not registered', async (t) => {
+  const testCfg = {
+    sockets: [
+      {
+        default: true,
+        name: 'test',
+        url: 'http://localhost:3000',
+        vuex: {
+          emitBacks: ['examples/sample', { 'examples/sample2': 'sample2' }]
+        }
+      }
+    ]
+  }
+  pOptions.set(testCfg)
+  const callCnt = { storeWatch: 0 }
+  const loadCnt = 2
+  for (let i = 0; i < loadCnt; i++) {
+    await loadPlugin({
+      t,
+      ioOpts: {
+        name: 'test',
+        channel: '/index'
+      },
+      callCnt
+    })
+  }
+
+  t.is(callCnt.storeWatch, 2)
+})
+
 test('Socket plugin (from nuxt.config)', async (t) => {
   delete require.cache[tmpFile]
   delete process.env.TEST
-  await compile(t)
-  const { ioServer } = t.context
-  const testSocket = await loadPlugin(t, {
-    name: 'test',
-    channel: '/index'
+  const imported = await compileAndImportPlugin({
+    src,
+    tmpFile,
+    options: io,
+    overwrite: true
+  }).catch((err) => {
+    console.error('Compile and Import err', err.message)
+    t.fail()
+  })
+
+  const testSocket = await loadPlugin({
+    t,
+    ioOpts: {
+      name: 'test',
+      channel: '/index'
+    },
+    plugin: imported.Plugin
   })
   const testJSON = { msg: 'it worked!' }
   const expected = 'It worked! Received msg: ' + JSON.stringify(testJSON)
+
   return new Promise((resolve) => {
-    testSocket
-      .emit('getMessage', testJSON, async (actual) => {
-        t.is(expected, actual)
-        await ioServer.stop()
-      })
-      .on('disconnect', () => {
-        resolve()
-      })
+    testSocket.emit('getMessage', testJSON, (actual) => {
+      t.is(expected, actual)
+      testSocket.close()
+      resolve()
+    })
   })
 })